+++ .env
@@ -8,3 +8,4 @@ REACT_APP_ORGANIZATION=/organization
 REACT_APP_SLIDES=/slides 
 REACT_APP_CATEGORIES=/categories
 REACT_APP_NEWS=/news
+REACT_APP_USERS=/users
