+++ src/Components/Backoffice/News/NewsList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../CardListStyles.css';
+import '../../CardListStyles.css';
 
 const NewsList = () => {
     const newsMock = [
