+++ src/Components/Backoffice/Projects/ProjectsForm.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import '../FormStyles.css';
+import '../../FormStyles.css';
 
 const ProjectsForm = () => {
   const [initialValues, setInitialValues] = useState({
