+++ src/Components/Backoffice/Slides/SlidesForm.js
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
-import '../FormStyles.css';
+import '../../FormStyles.css';
 
-const TestimonialForm = () => {
+const SlidesForm = () => {
     const [initialValues, setInitialValues] = useState({
-       name: '',
-       description: '' 
+        name: '',
+        description: ''
     });
 
     const handleChange = (e) => {
@@ -15,18 +15,18 @@ const TestimonialForm = () => {
         }
     }
 
-    const handleSubmit = (e) => {
+    const handleSubmit = (e) =>{
         e.preventDefault();
         console.log(initialValues);
     }
 
     return (
         <form className="form-container" onSubmit={handleSubmit}>
-            <input className="input-field" type="text" name="name" value={initialValues.name} onChange={handleChange} placeholder="Testimonial Title"></input>
-            <input className="input-field" type="text" name="description" value={initialValues.description} onChange={handleChange} placeholder="Testimonial description"></input>
+            <input className="input-field" type="text" name="name" value={initialValues.name} onChange={handleChange} placeholder="Slide Title"></input>
+            <input className="input-field" type="text" name="description" value={initialValues.description} onChange={handleChange} placeholder="Write the description"></input>
             <button className="submit-btn" type="submit">Send</button>
         </form>
     );
 }
  
-export default TestimonialForm;
+export default SlidesForm;
