+++ src/Components/Frontoffice/Activities/ActivitiesForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import '../FormStyles.css';
+import '../../FormStyles.css';
 
-const SlidesForm = () => {
+const ActivitiesForm = () => {
     const [initialValues, setInitialValues] = useState({
         name: '',
         description: ''
@@ -15,18 +15,18 @@ const SlidesForm = () => {
         }
     }
 
-    const handleSubmit = (e) =>{
+    const handleSubmit = (e) => {
         e.preventDefault();
         console.log(initialValues);
     }
-
+    
     return (
         <form className="form-container" onSubmit={handleSubmit}>
-            <input className="input-field" type="text" name="name" value={initialValues.name} onChange={handleChange} placeholder="Slide Title"></input>
-            <input className="input-field" type="text" name="description" value={initialValues.description} onChange={handleChange} placeholder="Write the description"></input>
+            <input className="input-field" type="text" name="name" value={initialValues.name} onChange={handleChange} placeholder="Activity Title"></input>
+            <input className="input-field" type="text" name="description" value={initialValues.description} onChange={handleChange} placeholder="Write some activity description"></input>
             <button className="submit-btn" type="submit">Send</button>
         </form>
     );
 }
  
-export default SlidesForm;
+export default ActivitiesForm;
