+++ src/Components/Frontoffice/Auth/RegisterForm.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import '../FormStyles.css';
+import '../../FormStyles.css';
 
 const RegisterForm = () => {
     const [initialValues, setInitialValues] = useState({
